requirements.txt
Django==1.9.6
django-countries==3.4.1
python-dateutil==2.5.3
requests==2.10.0
djangorestframework==3.3.3
django-crispy-forms==1.6.0
django-filter==0.13.0


# Tests
factory-boy==2.7.0
pytest==2.9.1
pytest-django==2.9.1
pytest-sugar==0.7.1
